\name{mcmc.core.inference.diagnostics}
\alias{mcmc.core.inference.diagnostics}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Run several diagnostics for the mcmc sampler.
}
\description{
This function makes convergence diagnostics for the mcmc sampler.
}
\usage{
mcmc.core.inference.diagnostics(mc, burnin)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
\item{mc}{
 a list containing information on MCMC sampler which is generated by functions 'mcmc.core.inference'.
}
\item{burnin}{
  the number of steps used for burn-in.
}
}

\examples{
# run examples in mcmc.core.inference first
mcmc.core.inference.diagnostics(mc, 2e3)
}
