\name{mcmc.diagnostics.approximation}
\alias{mcmc.diagnostics.approximation}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Run several diagnostics for the mcmc sampler.
}
\description{
This function makes convergence diagnostics for the mcmc sampler.
}
\usage{
mcmc.sarws.diagnostics(mc, burnin)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
\item{mc}{
 a list containing information on MCMC sampler which is generated by functions 'mcmc.sampling.approximation' and 'mcmc.sampling.approximation'.
}
\item{burnin}{
  the number of steps used for burn-in.
}
}

\examples{
# first run examples in 'mcmc.sampling.approximation' and 'mcmc.sampling.approximation'
mcmc.sarws.diagnostics(mc, 2e3)
}
